+++ src/main/java/com/commerce/core/common/config/security/service/CustomUserDetailsService.java
@@ -1,9 +1,9 @@
-package com.commerce.core.common.security.service;
+package com.commerce.core.common.config.security.service;
 
 import com.commerce.core.common.exception.CommerceException;
 import com.commerce.core.common.exception.ExceptionStatus;
-import com.commerce.core.common.security.vo.CustomUserDetails;
-import com.commerce.core.common.security.vo.JwtAuthentication;
+import com.commerce.core.common.config.security.vo.CustomUserDetails;
+import com.commerce.core.common.config.security.vo.JwtAuthentication;
 import com.commerce.core.member.entity.Member;
 import com.commerce.core.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
