+++ src/main/java/com/commerce/core/common/config/security/vo/AuthenticationInfo.java
@@ -1,0 +1,7 @@
+package com.commerce.core.common.config.security.vo;
+
+public interface AuthenticationInfo {
+    String getId();
+    Authority getAuthority();
+    String getName();
+}
