+++ src/main/java/com/commerce/core/common/config/security/vo/Authority.java
@@ -1,0 +1,5 @@
+package com.commerce.core.common.config.security.vo;
+
+public enum Authority {
+    ROLE_ADMIN, ROLE_USER
+}
