+++ src/main/java/com/commerce/core/common/config/security/vo/JwtIdentificationGenerateVO.java
@@ -1,4 +1,4 @@
-package com.commerce.core.common.security.vo;
+package com.commerce.core.common.config.security.vo;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
