--- src/main/java/com/commerce/core/common/security/IdentifierProvider.java
@@ -1,22 +1,0 @@
-package com.commerce.core.common.security;
-
-import com.commerce.core.common.security.vo.IdentificationGenerateVO;
-import org.springframework.security.core.Authentication;
-
-public interface IdentifierProvider {
-
-    /**
-     * Object For Authentication Generate
-     */
-    Object generateIdentificationInfo(IdentificationGenerateVO vo);
-
-    /**
-     * Authentication Info Get
-     */
-    Authentication getAuthenticationInfo(Object identificationInfo);
-
-    /**
-     * IdentificationInfo Valid Check
-     */
-    boolean validCheck(Object identificationInfo);
-}
