--- src/main/java/com/commerce/core/common/security/vo/AuthenticationInfo.java
@@ -1,7 +1,0 @@
-package com.commerce.core.common.security.vo;
-
-public interface AuthenticationInfo {
-    String getId();
-    Authority getAuthority();
-    String getName();
-}
