--- src/main/java/com/commerce/core/common/security/vo/Authority.java
@@ -1,5 +1,0 @@
-package com.commerce.core.common.security.vo;
-
-public enum Authority {
-    ROLE_ADMIN, ROLE_USER
-}
