--- src/main/java/com/commerce/core/common/security/vo/JwtToken.java
@@ -1,12 +1,0 @@
-package com.commerce.core.common.security.vo;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum JwtToken {
-    ACCESS_TOKEN(60 * 60 * 60 * 10L), REFRESH_TOKEN(60*60*1L);
-
-    private final Long expiredTime;
-}
