+++ src/main/java/com/commerce/core/member/service/LoginServiceImpl.java
@@ -2,10 +2,10 @@ package com.commerce.core.member.service;
 
 import com.commerce.core.common.exception.CommerceException;
 import com.commerce.core.common.exception.ExceptionStatus;
-import com.commerce.core.common.security.IdentifierProvider;
-import com.commerce.core.common.security.vo.IdentificationGenerateVO;
-import com.commerce.core.common.security.vo.JwtIdentificationGenerateVO;
-import com.commerce.core.common.security.vo.JwtToken;
+import com.commerce.core.common.config.security.IdentifierProvider;
+import com.commerce.core.common.config.security.vo.IdentificationGenerateVO;
+import com.commerce.core.common.config.security.vo.JwtIdentificationGenerateVO;
+import com.commerce.core.common.config.security.vo.JwtToken;
 import com.commerce.core.member.entity.Member;
 import com.commerce.core.member.vo.LoginDto;
 import com.commerce.core.member.vo.LoginSuccessDto;
